# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "XellPay API Documentation",
        "description": "xellpay api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://xellpay.test",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/v1/bill/calculate": {
            "post": {
                "tags": [
                    "Bill"
                ],
                "summary": "Calculate and store a bill for the subscriber",
                "operationId": "2238cd5cf4c3b1e7530eb7457c2dd889",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "month",
                                    "year"
                                ],
                                "properties": {
                                    "month": {
                                        "description": "Month to calculate bill for (1-12)",
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "year": {
                                        "description": "Year to calculate bill for",
                                        "type": "integer",
                                        "example": 2025
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bill calculated and stored.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bill successfully calculated!"
                                        },
                                        "data": {
                                            "properties": {
                                                "billDate": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-04-01"
                                                },
                                                "details": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "usageId": {
                                                                "type": "integer",
                                                                "example": 10
                                                            },
                                                            "amount": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 5
                                                            },
                                                            "usage": {
                                                                "properties": {
                                                                    "featureType": {
                                                                        "type": "string",
                                                                        "example": "CALL"
                                                                    },
                                                                    "usageAmount": {
                                                                        "type": "integer",
                                                                        "example": 500
                                                                    },
                                                                    "month": {
                                                                        "type": "integer",
                                                                        "example": 4
                                                                    },
                                                                    "year": {
                                                                        "type": "integer",
                                                                        "example": 2025
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Failed to calculate bill",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Bill calculation failed!"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bill/query": {
            "post": {
                "tags": [
                    "Bill"
                ],
                "summary": "Query total bill summary for a given subscriber, month, and year",
                "operationId": "65d21186783478d73e6707012967cb2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subscriber_no",
                                    "month",
                                    "year"
                                ],
                                "properties": {
                                    "subscriber_no": {
                                        "description": "Subscriber No",
                                        "type": "integer",
                                        "example": 10004567
                                    },
                                    "month": {
                                        "description": "Month to query the bill for (1-12)",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "year": {
                                        "description": "Year to query the bill for",
                                        "type": "integer",
                                        "example": 2025
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bill query successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Query successful."
                                        },
                                        "data": {
                                            "properties": {
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 60
                                                },
                                                "is_paid": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "bill_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-02-01"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Subscriber not found or no bill for that date",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No bill for date: February 2025"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer Token Authorization",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}